board: modular implementation, split out board_undo(), board_play() logic
board.c was doing too much, cleaner and faster this way:

board.c and board_undo.c now both include board_play.h
so each gets its own board_play() logic optimized for
its own needs, without code duplication.

for undo use board->u instead of passing undo pointer
through every function.

board.h cleanup
